Fix Windows compilation
This commit gets the Windows compilation working again, based on
all the hard work by Garrett. It also switches to fetching OpenFAST via
Git to the executable has version info. Lastly, it uses OpenBLAS instead
of the Netlib BLAS/LAPACK which should result in a performance boost.
The Choco install of MinGW seems to be unnecessary as the Windows
build host already contains MinGW64, though I'm not entirely sure
how this is maintained.